requirements.txt
kafka-python>=2.0.0
Flask>=2.0.0
